refactored code to move logic from LoginView to newly created LoginForm component
